requirements.txt
annotated-types==0.7.0
anyio==4.7.0
asyncio==3.4.3
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
fakeredis==2.26.2
fastapi==0.115.6
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pydantic==2.10.4
pydantic_core==2.27.2
pymongo==4.10.1
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-mock==3.14.0
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.41.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
